worker/requirements.txt
requests
radon
pygit2
gitpython